requirements.txt
python-telegram-bot
python-dotenv
requests
schedule
google-auth
google-genai
